Do not copy results to clipboard, if able to share
Previously, despite the share dialog, available on mobile devices, showing up, the results used to get copied to the clipboard as well regardless.

With this patch, the "Results copied to clipboard" text would not show up and the results would not be copied to the clipboard, if `navigator.share` is available, unless an error (excluding `AbortError`, fired when aborting the share) has occured.