Ability to play past puzzles (#10)
* Store data for multiple puzzles in an array

The `State::Puzzle` struct allows storing a puzzle's data separately from the main game state data.

Data for multiple puzzles can now be stored in an array. Data for multiple puzzles can be loaded/saved to `localStorage` by the JS layer.

Additionally, if a daily pool word list is missing, the chosen word is now truly random - `srand` is provided the current timestamp.

* Ability to play past puzzles

Logic and a GUI have been added to allow selecting past puzzles, based on ID.

Past puzzles are not counted towards statistics. Their intended use is to give the player a way to practice, or compete with others, despite having missed the day of the puzzle.

Additionally:

* Game input is no longer detected, when a popup window is active.
* Mutual properties and `visible(float T)` function between popup windows are now inherited from a base struct.

* Fix 2-space tabulation

* Remove workaround code for puzzle 643

The word was confirmed to have been fixed in the private daily pool.